# Native performance optimizations