## Summary

| Task | Metric | Before | After | Difference | % Change |
|------|--------|--------|-------|------------|----------|
| **Nested divs** | Latency avg (ns) | 30,725,126.06 | 27,848,582.47 | -2,876,543.59 | -9.36% |
| | Latency median (ns) | 30,348,498.00 | 27,827,932.50 | -2,520,565.50 | -8.30% |
| | Throughput avg (ops/s) | 33 | 36 | +3 | +9.09% |
| | Throughput median (ops/s) | 33 | 36 | +3 | +9.09% |
| **Sibling divs** | Latency avg (ns) | 30,095,190.50 | 27,493,222.28 | -2,601,968.22 | -8.65% |
| | Latency median (ns) | 29,846,615.00 | 27,508,688.00 | -2,337,927.00 | -7.83% |
| | Throughput avg (ops/s) | 33 | 36 | +3 | +9.09% |
| | Throughput median (ops/s) | 34 | 36 | +2 | +5.88% |

## React Native Fantom results

### Before

| (index) | Task name                                                             | Latency average (ns)  | Latency median (ns)      | Throughput average (ops/s) | Throughput median (ops/s) | Samples |
| ------- | --------------------------------------------------------------------- | --------------------- | ------------------------ | -------------------------- | ------------------------- | ------- |
| 0       | render 100 nested <html.div> with props and styles  | 30725126.06 ± 1.29% | 30348498.00 ± 19204.00 | 33 ± 1.19%               | 33                      | 64      |
| 1       | render 100 sibling <html.div> with props and styles | 30095190.50 ± 0.80% | 29846615.00 ± 6655.00  | 33 ± 0.78%               | 34                      | 64      |

### After

| (index) | Task name                                                             | Latency average (ns)  | Latency median (ns)      | Throughput average (ops/s) | Throughput median (ops/s) | Samples |
| ------- | --------------------------------------------------------------------- | --------------------- | ------------------------ | -------------------------- | ------------------------- | ------- |
| 0       | render 100 nested <html.div> with props and styles  | 27848582.47 ± 0.70% | 27827932.50 ± 4496.50  | 36 ± 0.69%               | 36                      | 64      |
| 1       | render 100 sibling <html.div> with props and styles | 27493222.28 ± 0.51% | 27508688.00 ± 25483.00 | 36 ± 0.50%               | 36                      | 64      |